requirements.txt
pyinstaller==6.12.0
typing-extensions==4.12.2
viam-sdk==0.41.0
opencv-python
pillow

pytest
pytest-asyncio
pylint

